+++ punic/errors.py
@@ -1,9 +1,7 @@
 from __future__ import division, absolute_import, print_function
 
 import contextlib
-
-from punic.logger import logger
-
+import logging
 
 class RepositoryNotClonedError(Exception):
     pass
@@ -29,17 +27,17 @@ def error_handling():
     try:
         yield
     except RepositoryNotClonedError:
-        logger.error('Error: No locally cloned repository found. Did you neglect to run `punic fetch` first?')
+        logging.error('Error: No locally cloned repository found. Did you neglect to run `punic fetch` first?')
         exit(-1)
     except CartfileNotFound as e:
-        logger.error('<err>Error</err>: No Cartfile found at path: <ref>{}</ref>'.format(e.path))
+        logging.error('<err>Error</err>: No Cartfile found at path: <ref>{}</ref>'.format(e.path))
         exit(-1)
     except NoSuchRevision as e:
-        logger.error('<err>Error</err>: No such revision {} found in repository {}'.format(e.revision, e.repository))
-        logger.error('Are you sure you are using the latest bits? Try an explicit `punic fetch` or use `punic bootstrap` instead of `punic build`')
+        logging.error('<err>Error</err>: No such revision {} found in repository {}'.format(e.revision, e.repository))
+        logging.error('Are you sure you are using the latest bits? Try an explicit `punic fetch` or use `punic bootstrap` instead of `punic build`')
         exit(-1)
     except PunicRepresentableError as e:
-        logger.error(e.message)
+        logging.error(e.message)
         exit(-1)
     except:
         raise