+++ punic/repository.py
@@ -7,11 +7,18 @@ import functools
 import hashlib
 import affirm
 import six
+import logging
+
+# Ideally we could six.urllib but this causes problem with nosetests!
+if six.PY2:
+    import urlparse
+elif six.PY3:
+    import urllib.parse as urlparse
+
 from memoize import mproperty
 from .runner import *
 from .config import *
 from .errors import *
-from .logger import *
 from .cartfile import *
 from .semantic_version import *
 
@@ -23,7 +30,8 @@ class Repository(object):
         if repo_path:
             self.path = repo_path
         else:
-            url_hash = hashlib.md5(self.identifier.remote_url).hexdigest()
+            remote_url = self.identifier.remote_url.encode("utf-8")
+            url_hash = hashlib.md5(remote_url).hexdigest()
             self.path = punic.config.repo_cache_directory / "{}_{}".format(self.identifier.project_name, url_hash)
 
         self.specifications_cache = dict()
@@ -77,7 +85,7 @@ class Repository(object):
 
     def checkout(self, revision):
         # type: (Revision)
-        logger.debug('Checking out <ref>{}</ref> @ revision <rev>{}</rev>'.format(self, revision))
+        logging.debug('Checking out <ref>{}</ref> @ revision <rev>{}</rev>'.format(self, revision))
         self.check_work_directory()
         try:
             runner.check_run('git checkout "{}"'.format(revision.sha), cwd=self.path)
@@ -86,10 +94,10 @@ class Repository(object):
 
     def fetch(self):
         if not self.path.exists():
-            logger.debug('<sub>Cloning</sub>: <ref>{}</ref>'.format(self))
+            logging.debug('<sub>Cloning</sub>: <ref>{}</ref>'.format(self))
 
             url = self.identifier.remote_url
-            import urlparse
+
             parsed_url = urlparse.urlparse(url)
             if parsed_url.scheme == 'file':
                 repo = parsed_url.path
@@ -100,7 +108,7 @@ class Repository(object):
         else:
             self.check_work_directory()
 
-            logger.info('<sub>Fetching</sub>: <ref>{}</ref>'.format(self))
+            logging.info('<sub>Fetching</sub>: <ref>{}</ref>'.format(self))
             runner.check_run('git fetch', cwd=self.path)
 
     def specifications_for_revision(self, revision):
@@ -186,7 +194,7 @@ class Revision(object):
             else:
                 return self.sha == other.sha
         except:
-            logger.error(self.repository)
+            logging.error(self.repository)
             raise
 
     # see: https://bugs.python.org/issue25732