+++ VERSION_HISTORY.markdown
@@ -1,5 +1,14 @@
 # Version History
 
+## 0.1.18
+
+* Make sub-project Build symlinks relative (fixes #11)
+* `punic init` is back and still as broken as ever.
+* Large logging cleanup and bug fix - upshot is `--no-color` works again
+* Adds a full `punic update` to test suite.
+* Fixes punic on Python 3.5 which was a little bit broken.
+* Fixes #8 - show a useful error if user doesn't have graphviz and/or pydotplus installed.
+
 ## 0.1.17
 
 * Really fixes problems with ReactiveCocoa